requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
einops==0.8.0
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.4
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pillow==10.2.0
PyYAML==6.0.2
requests==2.28.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sympy==1.12
threadpoolctl==3.5.0
timm==1.0.9
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==1.26.13
x-transformers==1.34.0
